Moved getAmplitudeAndFreq func in utils
